requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.6.20
Django @ file:///tmp/build/80754af9/django_1605066529526/work
psycopg2 @ file:///opt/concourse/worker/volumes/live/c8a4b965-959f-4e6c-7792-4aa881d63c8d/volume/psycopg2_1594305121341/work
python-decouple==3.3
pytz==2020.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
